debug

+++ Functions_Industry/js/industry.js
@@ -9,7 +9,7 @@
 var circCirc = function () {
 	prompt("What is the circumference of the circle (as a number, in inches, with a decimal. Example: 10.0 inches)");
 	
-	if (circCirc == "" || !parseInt(circCirc)){
+	if (circCirc == "" || !parseFloat(circCirc)){
 		alert("Something isn't right there.. Please enter your information again.");
 		circCirc();
 	} else {
